+++ client/src/App.js
@@ -1,13 +1,29 @@
-import logo from './logo.svg';
-import './App.css';
 import { Outlet } from 'react-router-dom';
+import { Container, Navbar } from 'react-bootstrap'; 
+import { LinkContainer } from 'react-router-bootstrap'; 
+
+import './App.css';
 
 function App() {
   return (
-     <> 
-      <Outlet/>
-     </>
+    <>
+    <Navbar bg='light' expand='lg' variant='light'>
+      <Container>
+        <LinkContainer to='/'>
+          <Navbar.Brand className='logo'>Taxi</Navbar.Brand>
+        </LinkContainer>
+        <Navbar.Toggle />
+        <Navbar.Collapse />
+      </Container>
+    </Navbar>
+    <Container className='pt-3'>
+      <Outlet />
+    </Container>
+  </>
   );
 }
+
+
+
 
 export default App;