+++ client/src/components/LogIn.js
@@ -1,16 +1,66 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
+import React, { useState } from "react";
+import { Link, Navigate } from "react-router-dom";
+import { Formik } from "formik";
+import { Breadcrumb, Button, Card, Form } from "react-bootstrap";
 
-function Landing (props) {
+function LogIn(props) {
+  const [isSubmitted, setSubmitted] = useState(false);
+  const onSubmit = (values, actions) => setSubmitted(true);
+
+  if (isSubmitted) {
+    return <Navigate to="/" />;
+  }
   return (
-    <div>
-      <h1>Taxi</h1>
-      <Link to='/'>Home</Link>
-      <br/>
-      <Link to='/sign-up'>Sign up</Link>
-      <Link to='/log-in'>Log in</Link>
-    </div>
+    <>
+      <Breadcrumb>
+        <Breadcrumb.Item href="/#/">Home</Breadcrumb.Item>
+        <Breadcrumb.Item active>Log in</Breadcrumb.Item>
+      </Breadcrumb>
+      <Card>
+        <Card.Header>Log in</Card.Header>
+        <Card.Body>
+          <Formik
+            initialValues={{
+              username: "",
+              password: "",
+            }}
+            onSubmit={onSubmit}
+          >
+            {({ handleChange, handleSubmit, values }) => (
+              <Form noValidate onSubmit={handleSubmit}>
+                <Form.Group className="mb-3" controlId="username">
+                  <Form.Label>Username:</Form.Label>
+                  <Form.Control
+                    name="username"
+                    onChange={handleChange}
+                    value={values.username}
+                  />
+                </Form.Group>
+                <Form.Group className="mb-3" controlId="password">
+                  <Form.Label>Password:</Form.Label>
+                  <Form.Control
+                    name="password"
+                    onChange={handleChange}
+                    type="password"
+                    value={values.password}
+                  />
+                </Form.Group>
+                <div className="d-grid mb-3">
+                  <Button type="submit" variant="primary">
+                    Log in
+                  </Button>
+                </div>
+              </Form>
+            )}
+          </Formik>
+
+          <Card.Text className="text-center">
+            Don't have an account? <Link to="/sign-up">Sign up!</Link>
+          </Card.Text>
+        </Card.Body>
+      </Card>
+    </>
   );
 }
 
-export default Landing;
+export default LogIn;