+++ client/src/index.js
@@ -1,3 +1,5 @@
+import 'bootswatch/dist/lumen/bootstrap.css';
+
 import React from 'react';
 import {createRoot} from 'react-dom/client';
 import { HashRouter, Route, Routes } from 'react-router-dom';